+++ public/release-notes.md
@@ -7,11 +7,14 @@ ________________________________________________________________________________
 #### Features
 - Authentication methods for Youtube and Twitch added to the menus
   - Application manages refresh tokens as well
+  - YouTube will Refresh tokens while app is live. Google access tokens expire after one hour requiring sign in.
+  - Check marks on main window indicated authentication status 
 - New chat command !odds added. This command will send a message back to the viewer informing them of the odds of winning
   - Possible Messages:
     - @user your odds of winning are xx.xx% and you have x entries
     - @user your odds of winning are xx.xx% and you have x entries you can double your odds by typing !here
     - @user you have not entered the wheel yet with !wheel
+- Message box will now show the service the viewer is on
 
 
 ___________________________________________________________________________________________